Circular footage work, but not shrinking
problems with taking footage at the top of the flight are fixed by
removing a lot of serial prints, was a memory full error. Attempted to
shrink normal Camera2SD with other Sd libraries, but then JPEG files are
not saved properly, header and footer are not present and file is much
shorter than should be. Changing libraraies might not save much in the
end anyways